requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
pydantic==1.10.7
sqlalchemy==1.4.46
pymysql==1.0.2
alembic==1.9.2
kafka-python==2.0.2
pytest==7.3.1
pytest-cov==4.0.0
python-dotenv==1.0.0